consumer/requirement.txt
confluent-kafka==2.10.0
Faker==37.1.0
numpy==2.2.5
psycopg2-binary==2.9.10
python-dotenv==1.1.0
tzdata==2025.2
